refactor(canon): split virtual_ts/expressions into cohesive modules
